# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Wizardry AI - Enterprise Deepfake Detection Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ─── Scans ───────────────────────────────────────────────────────────────────
...
  /subscriptions/upgrade:
    post:
      operationId: upgradeSubscription
      tags: [subscriptions]
      summary: Upgrade subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradeInput"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
  /subscriptions/tokens/consume:
    post:
      operationId: consumeTokens
      tags: [subscriptions]
      summary: Consume tokens for a scan operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenConsumeInput"
      responses:
        "200":
          description: Updated token balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenBalance"

  # ─── Settings ────────────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List API keys (Enterprise only)
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Generate a new API key (Enterprise only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: Created API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
  /api-keys/{id}:
    delete:
      operationId: deleteApiKey
      tags: [api-keys]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked
  /webhooks:
    get:
      operationId: listWebhooks
      tags: [api-keys]
      summary: List webhooks (Enterprise only)
      responses:
        "200":
          description: Webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
    post:
      operationId: createWebhook
      tags: [api-keys]
      summary: Create a webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookInput"
      responses:
        "201":
          description: Created webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
  /webhooks/{id}:
    delete:
      operationId: deleteWebhook
      tags: [api-keys]
      summary: Delete a webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Dashboard Stats ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, username, tier, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        tier:
          type: string
          enum: [free, basic, pro, advanced, enterprise]
        totalScans:
          type: integer
        aiDetected:
          type: integer
        authentic:
          type: integer
        uncertain:
          type: integer
        createdAt:
          type: string
    UserRegister:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
        username:
          type: string
        password:
          type: string
        displayName:
          type: string
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ─── Scan ───────────────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      required: [id, userId, planId, status, currentPeriodEnd]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: string
        status:
          type: string
          enum: [active, cancelled, past_due, trialing]
        currentPeriodEnd:
          type: string
        cancelAtPeriodEnd:
          type: boolean
    UpgradeInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
        paymentMethod:
          type: string
    TokenBalance:
      type: object
      required: [userId, planId, tokenLimit, tokensUsed, tokensRemaining]
      properties:
        userId:
          type: integer
        planId:
          type: string
        tokenLimit:
          type: ["integer", "null"]
        tokensUsed:
          type: integer
        tokensRemaining:
          type: ["integer", "null"]
        resetAt:
          type: ["string", "null"]
        warningThreshold:
          type: integer
        isExhausted:
          type: boolean
        isNearLimit:
          type: boolean
    TokenConsumeInput:
      type: object
      required: [scanId, amount]
      properties:
        scanId:
          type: integer
        amount:
          type: integer

    # ─── Settings ───────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, userId, type, title, message, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [scan_complete, token_warning, token_exhausted, upgrade_prompt, security_alert, greeting, reminder, suggestion, system]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high, critical]
        actionUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── API Keys ────────────────────────────────────────────────────────────────
    ApiKey:
      type: object
      required: [id, userId, name, keyPreview, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        keyPreview:
          type: string
        fullKey:
          type: ["string", "null"]
        active:
          type: boolean
        lastUsedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ApiKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    Webhook:
      type: object
      required: [id, userId, url, events, active, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        url:
          type: string
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        secret:
          type: ["string", "null"]
        createdAt:
          type: string
    WebhookInput:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
        events:
          type: array
          items:
            type: string

    # ─── Stats ───────────────────────────────────────────────────────────────────
...