# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wizardry AI - Enterprise Deepfake Detection Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserTier } from './userTier';

export interface User {
  id: number;
  email: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  tier: UserTier;
  totalScans?: number;
  aiDetected?: number;
  authentic?: number;
  uncertain?: number;
  createdAt: string;
}